Rename first-world.cpp to game.cpp
