+++ src/classes/game.cpp
@@ -5,7 +5,7 @@
 #include <string>
 #include "raylib.h"
 
-#include "first-world.h"
+#include "game.h"
 #include "Block.h"
 #include "World.h"
